// Front-End/src/app/models/profil.model.ts
export class User {

    id?: object;
    pseudo?: string;
    email?: string;
    password?: string;
    prenom?: string;
    nom?: string;
    genre?: string;
    birthDate?: string;
    rue?: string;
    codePostal?: string;
    ville?: string;
    photo?: string;
    presentation?: string;
    preferences?: string;
    amis?: string[];
    createdAt?: Date;

    // constructor(public pseudo: string,
    //             public email: string,
    //             public password: string,
    //             public prenom: string,
    //             public nom: string,
    //             public motDePasse: string,
    //             public genre?: string,
    //             public dateDeNaissance?: string,
    //             public rue?: string,
    //             public codePostal?: string,
    //             public ville?: string,
    //             public photo?: string,
    //             public presentation?: string,
    //             public preferences?: string,
    //             public amis?: string[],
    //             public dateCreationProfil?: string,
    //             ) {}

}